+++ djmoney/contrib/exchange/backends/__init__.py
@@ -1,0 +1,1 @@
+from .openexchangerates import OpenExchangeRatesBackend  # noqa