+++ djmoney/settings.py
@@ -25,3 +25,8 @@ if CURRENCY_CHOICES is None:
 
 CURRENCY_CHOICES.sort(key=operator.itemgetter(1, 0))
 DECIMAL_PLACES = getattr(settings, 'CURRENCY_DECIMAL_PLACES', 2)
+
+OPEN_EXCHANGE_RATES_URL = getattr(settings, 'OPEN_EXCHANGE_RATES_URL', 'https://openexchangerates.org/api/latest.json')
+OPEN_EXCHANGE_RATES_APP_ID = getattr(settings, 'OPEN_EXCHANGE_RATES_APP_ID', None)
+BASE_CURRENCY = getattr(settings, 'BASE_CURRENCY', 'USD')
+EXCHANGE_BACKEND = getattr(settings, 'EXCHANGE_BACKEND', 'djmoney.contrib.exchange.backends.OpenExchangeRatesBackend')